Adds config option to disable visual playlist
This adds the ability to disable the visual playlist by setting "visualplaylist" to false in the config of the player.  Visual playlist will no be included on the page when this is true.
Also modifies the styling of the next, previous, and playlist buttons so that it will automatically handle the spacing between the elements if the playlist icon is removed.

[Finishes #91596012]